일부 함수 분리작업 진행
